refactor(api): api/usage 를 도메인 모듈 5종으로 분할
1,244줄 → models 155 · jsonl 160 · anthropic 249 · codex 335 ·
routes 435 · __init__ 14. 전부 800 이내.

정의 이동만 했고 본문은 한 글자도 바꾸지 않았다 (split_audit 으로
정의 43개 · 모듈레벨 할당 19개 전부 본문 일치 확인, ruff format 이후
재실행해도 0건).

## 라우트를 한 모듈에 모은 이유

라우트 7개를 routes.py 하나에 원본 선언 순서 그대로 둔다. include_router
조립을 피하므로 등록 순서가 완전히 보존되고, 경로 가림 계약을 새로
만들지 않는다 (fastapi_include_order_is_contract 회피).

## 하중 지지대 — 테스트의 재바인딩까지 봐야 한다

소스만 보면 _codex_plan_cache 는 첨자 대입뿐이라(global 문 없음) 안전해
보이지만, 테스트가 monkeypatch.setattr 로 dict 를 통째로 갈아끼운다
(test_usage_jsonl.py:412·480·501). 재바인딩이 개입하면 읽는 쪽이 갈릴 때
한쪽이 옛 dict 를 계속 본다 — 그래서 캐시·헬퍼·라우트를 routes.py 에
함께 뒀다. _usage_cache 는 읽는 함수 둘(_load/_save)이 anthropic.py 에
함께 있어 분열하지 않는다.

## __init__.py 재노출을 좁게 한 이유

router 하나만 재노출한다 (실측: 외부 소비자는 api/app.py:89 뿐).
'원본 공개 표면 전체 재노출'을 따르면 __init__ 이 이동한 이름의 별칭을
만들어, monkeypatch.setattr(usage_mod, ...) 가 별칭만 갈아끼우고 정작
그 이름을 읽는 서브모듈은 원본을 계속 본다 — 테스트가 실제 홈
디렉토리를 스캔한 채 조용히 통과한다. 좁게 두면 AttributeError 로 즉시
드러나며, 실제로 갱신 대상 18곳이 전부 시끄럽게 실패했다.

## 테스트 갱신 (41건)

패치 타깃은 '그 이름을 읽는 함수가 사는 모듈'을 겨냥해야 한다 — 함수의
전역 조회는 자기 모듈 __dict__ 를 보기 때문이다. AST 로 읽는 함수를
실측해 매핑했다. test_usage_fetch_retry.py 는 참조가 전부 anthropic
소유라 import 한 줄로 정리했다.

Red-Green 검증: 27행 타깃을 분할 전 형태로 되돌리면 8 errors,
복원하면 16 passed.

게이트: ruff check 0 · ruff format 0 · mypy 304파일 0 ·
pytest 1 failed, 1360 passed, 2 skipped
(베이스라인 1357 + 신규 route_table 3건, 실패 1건은 알려진 RAG 플레이크)
